fix: resolve memory leak in CachedImage
Currently, the CachedImage class is repeatedly constructed by the
Transformer service. This would be okay, except that the constructor for
CacheImage called getLogger(), which ultimately invokes createNewDebug
inside the "debug" package. This creates new debugger instance which is
kept in a list of instances by the "debug" package and is thus never
removed. Consequently, everytime an image is transformed there is a
small leak for the new debug instance.

It looked like CachedImage was trying to be a singleton service. It
appeared that it was not such a service because Transformer took the
ImageAdapter as a function level parameter and needed to pass into the
CachedImage. I changed CachedImage to also take ImageAdapter as a
parameter and CachedImage into a singleton service thus avoiding the
constant reconstruction and memory leak.